requirements.txt
Django>=4.0,<5.0
slack_sdk
google-generativeai
# Add any other dependencies below